+++ Assets/Mirage/Samples~/Tanks/Scripts/TankGameManager.cs
@@ -97,6 +97,8 @@ namespace Mirage.Examples.Tanks
 
         bool GetAllReadyState()
         {
+            if (!LocalPlayer || !LocalPlayer.isReady) return false;
+            
             bool AllReady = true;
             foreach (Tank tank in players)
             {