+++ Assets/Mirage/Weaver/Processors/PropertySiteProcessor.cs
@@ -34,9 +34,19 @@ namespace Mirage.Weaver
             // does it set a field that we replaced?
             if (Setters.TryGetValue(opField, out MethodDefinition replacement))
             {
-                //replace with property
-                i.OpCode = OpCodes.Call;
-                i.Operand = replacement;
+                if (opField.DeclaringType.IsGenericInstance || opField.DeclaringType.HasGenericParameters) // We're calling to a generic class
+                {
+                    FieldReference newField = i.Operand as FieldReference;
+                    GenericInstanceType genericType = (GenericInstanceType)newField.DeclaringType;
+                    i.OpCode = OpCodes.Callvirt;
+                    i.Operand = replacement.MakeHostInstanceGeneric(genericType);
+                }
+                else
+                {
+                    //replace with property
+                    i.OpCode = OpCodes.Call;
+                    i.Operand = replacement;
+                }
             }
         }
 
@@ -46,9 +56,19 @@ namespace Mirage.Weaver
             // does it set a field that we replaced?
             if (Getters.TryGetValue(opField, out MethodDefinition replacement))
             {
-                //replace with property
-                i.OpCode = OpCodes.Call;
-                i.Operand = replacement;
+                if (opField.DeclaringType.IsGenericInstance || opField.DeclaringType.HasGenericParameters) // We're calling to a generic class
+                {
+                    FieldReference newField = i.Operand as FieldReference;
+                    GenericInstanceType genericType = (GenericInstanceType)newField.DeclaringType;
+                    i.OpCode = OpCodes.Callvirt;
+                    i.Operand = replacement.MakeHostInstanceGeneric(genericType);
+                }
+                else
+                {
+                    //replace with property
+                    i.OpCode = OpCodes.Call;
+                    i.Operand = replacement;
+                }
             }
         }
 
@@ -56,21 +76,39 @@ namespace Mirage.Weaver
         {
             if (instr.OpCode == OpCodes.Stfld && instr.Operand is FieldReference opFieldst)
             {
+                FieldReference resolved = opFieldst.Resolve();
+                if (resolved == null)
+                {
+                    resolved = opFieldst.DeclaringType.Resolve().GetField(opFieldst.Name);
+                }
+
                 // this instruction sets the value of a field. cache the field reference.
-                ProcessInstructionSetterField(instr, opFieldst);
+                ProcessInstructionSetterField(instr, resolved);
             }
 
             if (instr.OpCode == OpCodes.Ldfld && instr.Operand is FieldReference opFieldld)
             {
+                FieldReference resolved = opFieldld.Resolve();
+                if (resolved == null)
+                {
+                    resolved = opFieldld.DeclaringType.Resolve().GetField(opFieldld.Name);
+                }
+
                 // this instruction gets the value of a field. cache the field reference.
-                ProcessInstructionGetterField(instr, opFieldld);
+                ProcessInstructionGetterField(instr, resolved);
             }
 
             if (instr.OpCode == OpCodes.Ldflda && instr.Operand is FieldReference opFieldlda)
             {
+                FieldReference resolved = opFieldlda.Resolve();
+                if (resolved == null)
+                {
+                    resolved = opFieldlda.DeclaringType.Resolve().GetField(opFieldlda.Name);
+                }
+
                 // loading a field by reference,  watch out for initobj instruction
                 // see https://github.com/vis2k/Mirror/issues/696
-                return ProcessInstructionLoadAddress(md, instr, opFieldlda);
+                return ProcessInstructionLoadAddress(md, instr, resolved);
             }
 
             return instr;